delisy/requirements.txt
beautifulsoup4==4.6.1
certifi==2018.4.16
chardet==3.0.4
diff-match-patch==20121119
Django==2.0.7
django-background-tasks==1.1.13
django-compat==1.0.15
django-import-export==1.0.1
et-xmlfile==1.0.1
idna==2.7
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.5
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
